pyproject.toml
[project]
name = "saas-ops-simulation"
version = "1.0.0"
description = "A B2B SaaS Operations simulation open environment for testing AI reasoning, resource allocation, and technical debt management."
authors = [
    { name = "Anshuman Sharma", email = "nimishas899@gmail.com" }
]
dependencies = [
    "fastapi>=0.103.0",
    "uvicorn>=0.23.0",
    "pydantic>=2.0.0",
    "httpx>=0.24.1",
    "groq>=0.9.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "openenv-core>=0.2.0"
]
requires-python = ">=3.10"

[project.scripts]
server = "server.app:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
